Added roles and made admin page
