+++ Resturant/AdminMainPage.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace Resturant
+{
+    public partial class AdminMainPage : Form
+    {
+        public AdminMainPage()
+        {
+            InitializeComponent();
+        }
+
+        private void TextBox1_Click(object sender, EventArgs e)
+        {
+
+        }
+    }
+}
