+++ Resturant/User.cs
@@ -11,8 +11,7 @@ namespace Resturant
     internal class User
     {
         public string username, password, email, role;
-        private static string connectionString = "Data Source=(LocalDB)\\MSSQLLocalDB;AttachDbFilename=C:\\Users\\j0eTh\\source\\repos\\Resturant\\Resturant\\Database1.mdf;Integrated Security=True";
-
+        private static string connectionString = "Data Source=(LocalDB)\\MSSQLLocalDB;AttachDbFilename=C:\\Users\\alial\\source\\repos\\Ali1767\\DeepSleepResturant\\Resturant\\Database1.mdf;Integrated Security=True;Connect Timeout=30";
         private User(string username, string password, string email, string role)
         {
             this.username = username;
